requirements.txt
pandas
numpy
requests
ollama
openai
anthropic
google-generativeai
scikit-learn
tiktoken
rich